Make incompatible-pointer-types into an error
In the newer version of nccl net structures, the function signature are
different which will raise a warning of type incompatible-pointer-types.
To make sure we support the right nccl net version, we make the warning
into an error so the plugin is not created if the function signatures do
not match the nccl structure.

Signed-off-by: Raafat Feki <rfeki@cornelisnetworks.com>